# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Employee and Branch Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Employee and Branch Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The name field is required"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "phone"
                },
                "issue": {
                  "type": "string",
                  "example": "must be a valid phone number"
                }
              }
            },
            "description": "Detailed validation errors (optional)"
          }
        }
      },
      "Employee": {
        "type": "object",
        "required": [
          "id",
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the employee",
            "example": "emp_123abc"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Full name of the employee",
            "example": "Jane Smith"
          },
          "position": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Job position/title of the employee",
            "example": "Software Developer"
          },
          "department": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Department where the employee works",
            "example": "Engineering"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Employee's email address",
            "example": "jane.smith@company.com"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 20,
            "description": "Employee's phone number",
            "example": "+1-555-123-4567"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch where the employee works",
            "example": "branch_456def"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the employee record was created",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the employee record was last updated",
            "example": "2024-01-20T14:45:00Z"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/employees": {
      "get": {
        "summary": "Retrieve a list of all employees",
        "tags": [
          "Employees"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "position",
                  "department",
                  "email",
                  "phone",
                  "branchId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "John Doe"
                  },
                  "position": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Software Developer"
                  },
                  "department": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Engineering"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@company.com"
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 7,
                    "maxLength": 20,
                    "example": "+1-555-123-4567"
                  },
                  "branchId": {
                    "type": "string",
                    "example": "branch_123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "summary": "Retrieve a specific employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved employee",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an existing employee's information",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "position",
                  "department",
                  "email",
                  "phone",
                  "branchId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "John Doe"
                  },
                  "position": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Senior Software Developer"
                  },
                  "department": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 50,
                    "example": "Engineering"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@company.com"
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 7,
                    "maxLength": 20,
                    "example": "+1-555-123-4567"
                  },
                  "branchId": {
                    "type": "string",
                    "example": "branch_123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employees/branch/{branchId}": {
      "get": {
        "summary": "Retrieve all employees from a specific branch",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved employees from the branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found or no employees in this branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employees/department/{department}": {
      "get": {
        "summary": "Retrieve all employees from a specific department",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "department",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The name of the department"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved employees from the department",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found or no employees in this department",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}